+++ tests/testthat/test-catalogs.R
@@ -1,8 +1,6 @@
 test_that("Catalogs", {
-  skip_if_not(
-    tidyBdE:::bde_check_access(),
-    "Skipping... BdE not reachable."
-  )
+  skip_on_cran()
+  skip_if_bde_offline()
 
   # Test load catalogs----
   expect_error(bde_catalog_load("aa"))