+++ tests/testthat/test-indicators.R
@@ -1,8 +1,6 @@
 test_that("Indicators", {
-  skip_if_not(
-    tidyBdE:::bde_check_access(),
-    "Skipping... BdE not reachable."
-  )
+  skip_on_cran()
+  skip_if_bde_offline()
 
   # Test indicators----
   expect_silent(bde_ind_gdp_var())